+++ packages/http/package.js
@@ -3,12 +3,13 @@ Package.describe({
   version: '1.0.11-winr.5'
 });
 
-Npm.depends({request: "2.47.0"});
+Npm.depends({request: "2.53.0"});
 
 Package.onUse(function (api) {
   api.use('underscore');
   api.use('url');
   api.export('HTTP');
+  api.export('HTTPInternals', 'server');
   api.addFiles('httpcall_common.js', ['client', 'server']);
   api.addFiles('httpcall_client.js', 'client');
   api.addFiles('httpcall_server.js', 'server');
@@ -21,6 +22,7 @@ Package.onTest(function (api) {
   api.use('random');
   api.use('jquery', 'client');
   api.use('http', ['client', 'server']);
+  api.use('tinytest');
   api.use('test-helpers', ['client', 'server']);
 
   api.addFiles('test_responder.js', 'server');