+++ packages/mongo/mongo_livedata_tests.js
@@ -3082,3 +3082,21 @@ Meteor.isServer && Tinytest.add(
     });
   }
 );
+
+Meteor.isServer && Tinytest.add("mongo-livedata - npm modules", function (test) {
+  // Make sure the version number looks like a version number.
+  test.matches(MongoInternals.NpmModules.mongodb.version, /^1\.(\d+)\.(\d+)/);
+  test.equal(typeof(MongoInternals.NpmModules.mongodb.module), 'function');
+  test.equal(typeof(MongoInternals.NpmModules.mongodb.module.connect),
+             'function');
+  test.equal(typeof(MongoInternals.NpmModules.mongodb.module.ObjectID),
+             'function');
+
+  var c = new Mongo.Collection(Random.id());
+  var rawCollection = c.rawCollection();
+  test.isTrue(rawCollection);
+  test.isTrue(rawCollection.findAndModify);
+  var rawDb = c.rawDatabase();
+  test.isTrue(rawDb);
+  test.isTrue(rawDb.admin);
+});